Make all basic model routes uniform for all models
Added model-controller.js, which has the basic logic that's common to a
bunch of actions on a model. The actual controllers for each model
(eg category.js, post.js, etc) can just specify how the information in
the actual model is read/written and the rest of the logic is handled by
model-controller.js.

This way, if there's some standard information that ought to be in every
response, or some standard error message that ought to be returned for a
bad request, all that logic can be put in model-controller.js.

Changes to be committed:
      modified:   controllers/category.js
      modified:   controllers/comment.js
      modified:   controllers/index.js
      new file:   controllers/model-controller.js
      modified:   controllers/post.js
      modified:   controllers/user.js