Add authentication in backend with jsonwebtokens.
New packages have been installed for doing authentication:

- jsonwebtokens
- express-jwt
- passport
- passport-jwt

For all model components, authentication has been added to any actions
which require someone to be logged in, which is all creation and
modification of resources, except for:

- categories
- creation of a user (so that a guest can register).

For all model components, the user of the API has to be the owning user
of a resource in order to delete or modify it, except for categories.

In order to authenticate, you must visit the `/login` route and provide
credentials in order to receive a token. From there on in, the token
represents who you are.

On branch login
Changes to be committed:
      new file:   controllers/authenticator.js
      modified:   controllers/category.js
      modified:   controllers/comment.js
      modified:   controllers/index.js
      new file:   controllers/login.js
      modified:   controllers/model-controller.js
      modified:   controllers/post.js
      modified:   controllers/user.js
      modified:   package-lock.json
      modified:   package.json